+++ pages/settings.tsx
@@ -3,8 +3,6 @@ import BackButton from "../components/BackButton/BackButton";
 import Topic from "../components/Topic/Topic";
 import styles from "../styles/settings.module.css";
 import Lottie from "react-lottie";
-// import animationData from "../lotties/loading.json";
-// import animationData from "../lotties/404.json";
 import animationData from "../lotties/construction.json";
 export default function Settings() {
   const defaultOptions = {
@@ -34,7 +32,7 @@ export default function Settings() {
       </header>
       <main className={styles.main}>
         <div className={styles.loading}>
-          <Lottie options={defaultOptions} height={220} width={300} />
+          <Lottie options={defaultOptions} height={300} width={300} />
         </div>
       </main>
       <footer className={styles.footer}></footer>
